% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{add_file_paths}
\alias{add_file_paths}
\title{Adds a standardized file_paths.R project root directory}
\usage{
add_file_paths(dir_R = here::here())
}
\description{
Adds a standardized file_paths.R project root directory
}
