% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{create_project_directories}
\alias{create_project_directories}
\title{Creates the standard set of subdirectories for the EDUC 3089 Course}
\usage{
create_project_directories()
}
\description{
This function sets up the data, script, analysis, and output folders,
and generates the necessary helper files (README, file_paths.R, .gitignore).
}
