% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_extra_packages}
\alias{install_extra_packages}
\title{Install Extra Course Packages}
\usage{
install_extra_packages()
}
\description{
Installs heavy or system-dependent packages (like pdftools, tesseract, text).
Run this function if you need PDF processing, OCR, or Python-based text models.
}
